DEF-193: all the things
