+++ packages/integrations/reducers/defaultState.js
@@ -1,6 +1,14 @@
+export const manager = {
+  componentMap: [],
+  hydrated: false,
+};
+
+export const pico = {};
+
+export const coral = {};
+
 export default {
-  integrations: {
-    componentMap: [],
-    hydrated: false,
-  },
+  manager,
+  pico,
+  coral,
 };
