+++ packages/integrations/reducers/integrationsReducer.js
@@ -1,18 +1,6 @@
 import merge from 'lodash/fp/merge';
-import {
-  HYDRATE_COMPONENTS,
-  RECEIVE_CORAL_SSO_TOKEN,
-  RECEIVE_CORAL_LOGOUT_REQUEST,
-  RECEIVE_CORAL_LOGOUT,
-  RECEIVE_CORAL_USERNAME_REQUEST,
-  SET_CORAL_USERNAME,
-  RECEIVE_CORAL_USERNAME_VALIDATION_ERROR,
-  REQUIRE_UPGRADE_FOR_CORAL_SSO,
-  RECEIVE_PICO_PLAN_UPGRADE,
-  PICO_LOADED,
-  UPDATE_PICO_PAGE_INFO,
-} from '../actions/types';
-import defaultState from './defaultState';
+import { HYDRATE_COMPONENTS } from '../actions/types';
+import { manager as defaultState } from './defaultState';
 
 /**
  * Handle Redux actions operating on the integrations state slice.
@@ -27,90 +15,6 @@ export default function integrationsReducer(
   switch (type) {
     case HYDRATE_COMPONENTS:
       return merge(state, { componentMap: payload, hydrated: true });
-
-    case PICO_LOADED:
-      return {
-        ...state,
-        pico: {
-          ...state.pico,
-          loaded: true,
-        },
-      };
-
-    case UPDATE_PICO_PAGE_INFO:
-      return {
-        ...state,
-        pico: {
-          ...state.pico,
-          pageInfo: payload,
-        },
-      };
-
-    case REQUIRE_UPGRADE_FOR_CORAL_SSO:
-      return {
-        ...state,
-        coral: {
-          ...state.coral,
-          requireUpgrade: true,
-        },
-      };
-
-    case RECEIVE_PICO_PLAN_UPGRADE:
-      return {
-        ...state,
-        coral: {
-          ...state.coral,
-          requireUpgrade: false,
-        },
-      };
-
-    case RECEIVE_CORAL_USERNAME_REQUEST:
-      return {
-        ...state,
-        coral: {
-          ...state.coral,
-          requireUsername: true,
-        },
-      };
-
-    case SET_CORAL_USERNAME:
-      return {
-        ...state,
-        coral: {},
-      };
-
-    case RECEIVE_CORAL_USERNAME_VALIDATION_ERROR:
-      return {
-        ...state,
-        coral: {
-          ...state.coral,
-          validationError: payload,
-        },
-      };
-
-    case RECEIVE_CORAL_SSO_TOKEN:
-      return {
-        ...state,
-        coral: {
-          ...state.coral,
-          token: payload,
-        },
-      };
-
-    case RECEIVE_CORAL_LOGOUT_REQUEST:
-      return {
-        ...state,
-        coral: {
-          token: null,
-          purgeUser: true,
-        },
-      };
-
-    case RECEIVE_CORAL_LOGOUT:
-      return {
-        ...state,
-        coral: {},
-      };
 
     default:
       return state;
